requirements.txt
django>=1.4.2
oauth2>=1.5.167
python_openid>=2.2
django-crispy-forms>=1.2.3

